requirements.txt
mssql-django 
django-bootstrap-v5
